Move redirectToStatsIfLocationConnected to controller
